Exchange the status labels "Draft" and "Exploratory".
The first official status in repository is now named "Exploratory". This fits
better both with its tagline "we think this idea is worth exploring" and also
how people have been using the term "draft" on the list.

Clarify that anyone competent (ie commit access) can accept an *Exploratory*
RFC - like PEPs, the intention is that we consider any reasonable idea, even
if seems likely to be rejected, as the reasoning for rejection is useful to
record.

(Obviously, once we have "institutional memory" of why certain ideas get
rejected, likely we can filter out similar ideas without needing an RFC,
but we're not there yet.)