update the PPC status to be more consistent
The different values for PPC status are somewhat inconsistent between the
various documents in the docs/ directory, and the actual PPC documents.

For example, "Accepted" in the process document means "with the code
merged as a new (probably experimental!) feature of perl", while in the
current PPCs documents, it was used to mean "Accepted Proposal".

During PSC #178 2025-02-06, Aristotle and BooK discussed the issue,
and proposed some new status names, with a goal of improving clarity.

This commit makes the following status renames:

    Draft       => Proposed
    Exploratory => Proposed
    Accepted    => Wanted
    Implemented => Incorporated

The discussion also revolved about the possible parallel existence of an
idea (the PPC proposal) and an implementation. An implementation can exist
while the PPC is still being discussed an evaluated, and help steer it.
An implementation can be rejected, while the idea is still accepted.